+++ examples/bare-metal/visual_wake_words/src/xtflm_conf.h
@@ -1,0 +1,17 @@
+// Copyright 2020-2021 XMOS LIMITED.
+// This Software is subject to the terms of the XMOS Public Licence: Version 1.
+
+#ifndef XTFLM_CONF_h_
+#define XTFLM_CONF_h_
+
+#define XTFLM_OPERATORS        (32)
+
+#define NETWORK_NUM_THREADS    (5)
+#define AISRV_GPIO_LENGTH      (4)
+
+#define NUM_OUTPUT_TENSORS     (1)
+#define NUM_INPUT_TENSORS      (1)
+
+#define MAX_DEBUG_LOG_LENGTH   (256)
+
+#endif // XTFLM_CONF_h_
